+++ tests/fixtures/next-pages/pages/_document.tsx
@@ -1,0 +1,3 @@
+export default function MyDocument() {
+  return null
+}